Factor ClientState into own module
